requirements.txt
requests
psycopg2-binary
pgvector
rich
pandas
pypdf
PyPDF2
pymupdf
pdfminer.six
